iox-#1391 Move files to time module
