+++ iceoryx_dust/include/iceoryx_dust/posix_wrapper/message_queue.hpp
@@ -19,10 +19,10 @@
 
 #include "iceoryx_dust/design/creation.hpp"
 #include "iceoryx_hoofs/internal/posix_wrapper/ipc_channel.hpp"
-#include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_platform/fcntl.hpp"
 #include "iceoryx_platform/mqueue.hpp"
 #include "iceoryx_platform/stat.hpp"
+#include "iox/duration.hpp"
 #include "iox/optional.hpp"
 
 namespace iox
