+++ iceoryx_dust/include/iceoryx_dust/posix_wrapper/named_pipe.hpp
@@ -20,9 +20,9 @@
 #include "iceoryx_hoofs/concurrent/lockfree_queue.hpp"
 #include "iceoryx_hoofs/internal/posix_wrapper/ipc_channel.hpp"
 #include "iceoryx_hoofs/internal/posix_wrapper/shared_memory_object.hpp"
-#include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_hoofs/posix_wrapper/unnamed_semaphore.hpp"
 #include "iceoryx_platform/semaphore.hpp"
+#include "iox/duration.hpp"
 #include "iox/expected.hpp"
 #include "iox/optional.hpp"
 #include "iox/string.hpp"
