+++ iceoryx_dust/source/posix_wrapper/named_pipe.cpp
@@ -16,7 +16,7 @@
 
 #include "iceoryx_dust/posix_wrapper/named_pipe.hpp"
 #include "iceoryx_dust/cxx/std_string_support.hpp"
-#include "iceoryx_hoofs/cxx/deadline_timer.hpp"
+#include "iox/deadline_timer.hpp"
 #include "iox/into.hpp"
 
 #include <thread>
