+++ iceoryx_hoofs/BUILD.bazel
@@ -36,7 +36,7 @@ cc_library(
         "memory/source/*.cpp",
         "design/source/*.cpp",
     ]),
-    hdrs = glob(["include/**"]) + glob(["legacy/**"]) + glob(["memory/**"]) + glob(["container/**"]) + glob(["vocabulary/**"]) + glob(["utility/**"]) + glob(["primitives/**"]) + glob(["design/**"]) + [
+    hdrs = glob(["include/**"]) + glob(["legacy/**"]) + glob(["memory/**"]) + glob(["container/**"]) + glob(["vocabulary/**"]) + glob(["time/**"]) + +glob(["utility/**"]) + glob(["primitives/**"]) + glob(["posix/**"]) + glob(["design/**"]) + [
         ":iceoryx_hoofs_deployment_hpp",
     ],
     includes = [
@@ -45,7 +45,9 @@ cc_library(
         "include/",
         "legacy/include/",
         "memory/include/",
+        "posix/time/include/",
         "primitives/include/",
+        "time/include/",
         "utility/include/",
         "vocabulary/include/",
     ],
