+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/posix_wrapper/semaphore_interface.hpp
@@ -16,8 +16,8 @@
 #ifndef IOX_HOOFS_POSIX_WRAPPER_SEMAPHORE_INTERFACE_HPP
 #define IOX_HOOFS_POSIX_WRAPPER_SEMAPHORE_INTERFACE_HPP
 
-#include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_platform/semaphore.hpp"
+#include "iox/duration.hpp"
 #include "iox/expected.hpp"
 
 namespace iox
