+++ iceoryx_hoofs/test/moduletests/test_posix_semaphore_interface.cpp
@@ -16,7 +16,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "iceoryx_hoofs/internal/posix_wrapper/semaphore_interface.hpp"
-#include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_hoofs/posix_wrapper/named_semaphore.hpp"
 #include "iceoryx_hoofs/posix_wrapper/unnamed_semaphore.hpp"
 #include "iceoryx_hoofs/testing/test.hpp"
@@ -24,6 +23,7 @@
 #include "iceoryx_hoofs/testing/watch_dog.hpp"
 #include "iceoryx_platform/platform_settings.hpp"
 #include "iceoryx_platform/time.hpp"
+#include "iox/duration.hpp"
 
 #include "test.hpp"
 #include "test_posix_semaphore_common.hpp"
