+++ iceoryx_hoofs/test/moduletests/test_time_adaptive_wait.cpp
@@ -17,12 +17,12 @@
 #include "test.hpp"
 using namespace ::testing;
 
-#include "iceoryx_hoofs/internal/cxx/adaptive_wait.hpp"
+#include "iox/detail/adaptive_wait.hpp"
 
 #include <atomic>
 #include <thread>
 
-using namespace iox::internal;
+using namespace iox::detail;
 
 namespace
 {
