+++ iceoryx_hoofs/test/moduletests/test_time_deadline_timer.cpp
@@ -15,9 +15,9 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/cxx/deadline_timer.hpp"
-#include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_hoofs/testing/timing_test.hpp"
+#include "iox/deadline_timer.hpp"
+#include "iox/duration.hpp"
 #include "test.hpp"
 
 #include <chrono>
