+++ iceoryx_hoofs/test/moduletests/test_time_unit_duration.cpp
@@ -15,10 +15,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_hoofs/posix_wrapper/posix_call.hpp"
 #include "iceoryx_hoofs/testing/mocks/logger_mock.hpp"
 #include "iceoryx_hoofs/testing/testing_logger.hpp"
+#include "iox/duration.hpp"
 #include "test.hpp"
 #include <ctime>
 #include <iostream>
