+++ iceoryx_posh/include/iceoryx_posh/internal/popo/building_blocks/chunk_distributor.inl
@@ -169,7 +169,7 @@ inline uint64_t ChunkDistributor<ChunkDistributorDataType>::deliverToAllStoredQu
     }
 
     // busy waiting until every queue is served
-    iox::internal::adaptive_wait adaptiveWait;
+    iox::detail::adaptive_wait adaptiveWait;
     while (!remainingQueues.empty())
     {
         adaptiveWait.wait();
