+++ iceoryx_posh/test/integrationtests/test_mq_interface_startup_race.cpp
@@ -19,10 +19,10 @@
 #include "test.hpp"
 
 #include "iceoryx_dust/posix_wrapper/message_queue.hpp"
-#include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_hoofs/posix_wrapper/posix_call.hpp"
 #include "iceoryx_posh/internal/runtime/ipc_message.hpp"
 #include "iceoryx_posh/internal/runtime/ipc_runtime_interface.hpp"
+#include "iox/duration.hpp"
 
 
 #include <chrono>
