+++ iceoryx_posh/test/moduletests/test_popo_user_trigger.cpp
@@ -15,9 +15,9 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_posh/iceoryx_posh_types.hpp"
 #include "iceoryx_posh/popo/user_trigger.hpp"
+#include "iox/duration.hpp"
 
 #include "test.hpp"
